Avoid duplicating the files in memory.
Run gofmt on the source without file data,
then inject the data using fasttemplate, which allows
writing directly to the output instead of passing around strings.

There's an empty line before the Content key in EmbeddedFile
because gofmt does not align the value with previous line
if the value is very long.
Since gofmt doesn't see the actual value, it can't align it correctly.