goimport-friendly template.
This stops goimport from persistently trying to reformat this file by grouping imports the way it expects. We use a combination of `goimport` and `rice embed-go` to embed data in binaries, and our CI "check that rice-box.go files are up to date" script causes goimport errors; running `goimport`, in turn, causes the files to be modified back the next time we run `rice`.